Allow to paste in editable list view
Paste is allowed only on editable list.
The paste starts at the cell selected or the first one if None.
The paste stops on a field if it can not set the value or if the resulted value
is different.
The paste set the focus on the last record modified and save each record one by
one.

review530002