Check if record is still in the multiselection when displaying
The records of current value could no more be in the selection if it has changed
because of the domain. So it must no more be displaied and it will be removed by
the call to set_client.

issue4029
review9541002
(grafted from aec021eae9589e0e0550810b26b710219c7eeeb1)

--HG--
branch : 3.2