Trigger editing-done on ComboBox changed only if a row is selected
The CellRendererTime uses also a ComboBoxEntry for editing but parsing it on
every change does not work because the parsing will expect a valid time format
otherwise the current text will be erased.
So to keep the original goal of changeset 3591350f0018 which is to trigger
"editing-done" when an option from the drop-down menu is selected, the callback
on "changed" must ensure there is an active row selected.

issue5746
review28601002