Ensure sort key always return an integer
When sorting tree model, it is possible that some ids are not in the ordered
list (because they do not fit anymore in the domain). So we return the number
of records so they are moved at the end of the list.

issue8042
review53531002
(grafted from 74557f917ae5ac2c96d45aa90b21019e54a7db74)

--HG--
branch : 4.8