Do not use record in __init__ of form widget
During the initialization of the widget, the view may not yet have a proper
record. But update_img access the record of the view.
Also it is safer for list-form to define the _record before calling super in
case there is an access to record attribute.

issue8616
review285981002
(grafted from 032754bdac9d67ad96ecbf17d3732edf3c2399af)

--HG--
branch : 5.2