Do not search remote value if view has no more the focus
If the user switch the tab with a value filled in the entry, the search window
is hidden and the all application is made insensitive. So we must ensure that
the search window will be linked to the current view. This is ensured by a
widget with the focus.

issue8733
review288291002