Catch any exception for server version RPC
It is the first method called with the input of the user.
The login window does not support that it raises an exception and as a lot of
different exceptions can be raised (and may change depending of Python
version), we must catch all of them. It does not hide any programming error as
they are logged.

issue9408
review292561004