Return false when modified is not saved and not created
The return of modified_save method is used to decide to execute an action or
not. When a new record is cancelled, the client select the previous record so
we must not run the action on this one.

issue9881
review337111002
(grafted from 392c4a9f154c8c7094d2dbadef53dec0b1ca2b01)

--HG--
branch : 5.6