Do not display search_count if it is not higher than limit
On small set of record when records are deleted the number of record could
become lower than the initial search_count. As maintaining a correct
search_count is complex, it is simpler to never display it if it is lower than
the search limit.

issue10789
review383501003
(grafted from 07165c9a8fc19768f45170df622e0998e85abee5)

--HG--
branch : 5.0